refactor: split and group related methods
