+++ pkg/rpaas/client/info.go
@@ -1,0 +1,52 @@
+// Copyright 2021 tsuru authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package client
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/tsuru/rpaas-operator/pkg/rpaas/client/types"
+)
+
+func (args InfoArgs) Validate() error {
+	if args.Instance == "" {
+		return ErrMissingInstance
+	}
+
+	return nil
+}
+
+func (c *client) Info(ctx context.Context, args InfoArgs) (*types.InstanceInfo, error) {
+	if err := args.Validate(); err != nil {
+		return nil, err
+	}
+
+	pathName := fmt.Sprintf("/resources/%s/info", args.Instance)
+	req, err := c.newRequest("GET", pathName, nil, args.Instance)
+	if err != nil {
+		return nil, err
+	}
+
+	response, err := c.do(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode != http.StatusOK {
+		return nil, newErrUnexpectedStatusCodeFromResponse(response)
+	}
+
+	var infoPayload types.InstanceInfo
+	err = json.NewDecoder(response.Body).Decode(&infoPayload)
+	if err != nil {
+		return nil, err
+	}
+
+	return &infoPayload, nil
+}
