+++ pkg/rpaas/client/internal_client.go
@@ -5,19 +5,14 @@
 package client
 
 import (
-	"bytes"
 	"context"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"mime/multipart"
 	"net/http"
-	"net/url"
 	"os"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/gorilla/websocket"
@@ -161,710 +156,6 @@ func (c *client) GetPlans(ctx context.Context, instance string) ([]types.Plan, e
 
 func (c *client) GetFlavors(ctx context.Context, instance string) ([]types.Flavor, error) {
 	return nil, fmt.Errorf("not implemented yet")
-}
-
-func (args ScaleArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.Replicas < int32(0) {
-		return fmt.Errorf("rpaasv2: replicas must be greater or equal than zero")
-	}
-
-	return nil
-}
-
-func (c *client) Scale(ctx context.Context, args ScaleArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("Scale", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args InfoArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	return nil
-}
-
-func (c *client) Info(ctx context.Context, args InfoArgs) (*types.InstanceInfo, error) {
-	if err := args.Validate(); err != nil {
-		return nil, err
-	}
-
-	request, err := c.buildRequest("Info", args)
-	if err != nil {
-		return nil, err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return nil, err
-	}
-	defer response.Body.Close()
-
-	if response.StatusCode != http.StatusOK {
-		return nil, newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	var infoPayload types.InstanceInfo
-	err = json.NewDecoder(response.Body).Decode(&infoPayload)
-	if err != nil {
-		return nil, err
-	}
-
-	return &infoPayload, nil
-}
-
-func (args UpdateCertificateArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.Certificate == "" {
-		return fmt.Errorf("rpaasv2: certificate cannot be empty")
-	}
-
-	if args.Key == "" {
-		return fmt.Errorf("rpaasv2: key cannot be empty")
-	}
-
-	return nil
-}
-
-func (c *client) UpdateCertificate(ctx context.Context, args UpdateCertificateArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("UpdateCertificate", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args *DeleteCertificateArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	return nil
-}
-
-func (c *client) DeleteCertificate(ctx context.Context, args DeleteCertificateArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	args.Name = url.QueryEscape(args.Name)
-	request, err := c.buildRequest("DeleteCertificate", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args UpdateBlockArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.Name == "" {
-		return ErrMissingBlockName
-	}
-
-	if args.Content == "" {
-		return fmt.Errorf("rpaasv2: content cannot be empty")
-	}
-
-	return nil
-}
-
-func (c *client) UpdateBlock(ctx context.Context, args UpdateBlockArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("UpdateBlock", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args DeleteBlockArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.Name == "" {
-		return ErrMissingBlockName
-	}
-
-	return nil
-}
-
-func (c *client) DeleteBlock(ctx context.Context, args DeleteBlockArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("DeleteBlock", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args ListBlocksArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	return nil
-}
-
-func (c *client) ListBlocks(ctx context.Context, args ListBlocksArgs) ([]types.Block, error) {
-	if err := args.Validate(); err != nil {
-		return nil, err
-	}
-
-	request, err := c.buildRequest("ListBlocks", args)
-	if err != nil {
-		return nil, err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return nil, err
-	}
-
-	defer response.Body.Close()
-	if response.StatusCode != http.StatusOK {
-		return nil, newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	var blockList struct {
-		Blocks []types.Block `json:"blocks"`
-	}
-	if err = unmarshalBody(response, &blockList); err != nil {
-		return nil, err
-	}
-
-	return blockList.Blocks, nil
-}
-
-func (args DeleteRouteArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.Path == "" {
-		return ErrMissingPath
-	}
-
-	return nil
-}
-
-func (c *client) DeleteRoute(ctx context.Context, args DeleteRouteArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("DeleteRoute", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	defer response.Body.Close()
-	if response.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args ListRoutesArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	return nil
-}
-
-func (c *client) ListRoutes(ctx context.Context, args ListRoutesArgs) ([]types.Route, error) {
-	if err := args.Validate(); err != nil {
-		return nil, err
-	}
-
-	request, err := c.buildRequest("ListRoutes", args)
-	if err != nil {
-		return nil, err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return nil, err
-	}
-
-	defer response.Body.Close()
-
-	if response.StatusCode != http.StatusOK {
-		return nil, newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	var routes struct {
-		Routes []types.Route `json:"paths"`
-	}
-	if err = unmarshalBody(response, &routes); err != nil {
-		return nil, err
-	}
-
-	return routes.Routes, nil
-}
-
-func (args UpdateRouteArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.Path == "" {
-		return ErrMissingPath
-	}
-
-	return nil
-}
-
-func (c *client) UpdateRoute(ctx context.Context, args UpdateRouteArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("UpdateRoute", args)
-	if err != nil {
-		return err
-	}
-
-	response, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	defer response.Body.Close()
-
-	if response.StatusCode != http.StatusCreated {
-		return newErrUnexpectedStatusCodeFromResponse(response)
-	}
-
-	return nil
-}
-
-func (args GetAutoscaleArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	return nil
-}
-
-func (c *client) GetAutoscale(ctx context.Context, args GetAutoscaleArgs) (*types.Autoscale, error) {
-	if err := args.Validate(); err != nil {
-		return nil, err
-	}
-
-	request, err := c.buildRequest("GetAutoscale", args)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := c.do(ctx, request)
-	if err != nil {
-		return nil, err
-	}
-
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, newErrUnexpectedStatusCodeFromResponse(resp)
-	}
-
-	defer resp.Body.Close()
-	var spec *types.Autoscale
-	err = json.NewDecoder(resp.Body).Decode(&spec)
-	if err != nil {
-		return nil, err
-	}
-
-	return spec, nil
-}
-
-func (args UpdateAutoscaleArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if args.MaxReplicas == 0 && args.MaxReplicas == args.MinReplicas && args.MaxReplicas == args.CPU && args.MaxReplicas == args.Memory {
-		return ErrMissingValues
-	}
-	return nil
-}
-
-func (c *client) shouldCreate(ctx context.Context, instance string) (bool, error) {
-	_, err := c.GetAutoscale(ctx, GetAutoscaleArgs{Instance: instance})
-	if err != nil {
-		if isNotFoundError(err) {
-			return true, nil
-		}
-		return false, err
-	}
-	return false, nil
-}
-
-func (c *client) UpdateAutoscale(ctx context.Context, args UpdateAutoscaleArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	var request *http.Request
-	shouldCreate, err := c.shouldCreate(ctx, args.Instance)
-	if err != nil {
-		return err
-	}
-	if shouldCreate {
-		request, err = c.buildRequest("CreateAutoscale", args)
-		if err != nil {
-			return err
-		}
-	} else {
-		request, err = c.buildRequest("UpdateAutoscale", args)
-		if err != nil {
-			return err
-		}
-	}
-
-	resp, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if resp.StatusCode != http.StatusCreated {
-		return newErrUnexpectedStatusCodeFromResponse(resp)
-	}
-
-	return nil
-}
-
-func (args RemoveAutoscaleArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-	return nil
-}
-
-func (c *client) RemoveAutoscale(ctx context.Context, args RemoveAutoscaleArgs) error {
-	if err := args.Validate(); err != nil {
-		return err
-	}
-
-	request, err := c.buildRequest("RemoveAutoscale", args)
-	if err != nil {
-		return err
-	}
-
-	resp, err := c.do(ctx, request)
-	if err != nil {
-		return err
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return newErrUnexpectedStatusCodeFromResponse(resp)
-	}
-
-	return nil
-}
-
-func (args ExecArgs) Validate() error {
-	if args.Instance == "" {
-		return ErrMissingInstance
-	}
-
-	if len(args.Command) == 0 {
-		return ErrMissingExecCommand
-	}
-
-	return nil
-}
-
-func (c *client) Exec(ctx context.Context, args ExecArgs) (*websocket.Conn, error) {
-	if err := args.Validate(); err != nil {
-		return nil, err
-	}
-
-	serverAddress := c.formatURL(fmt.Sprintf("/resources/%s/exec", args.Instance), args.Instance)
-	u, err := url.Parse(serverAddress)
-	if err != nil {
-		return nil, err
-	}
-
-	if u.Scheme == "https" {
-		u.Scheme = "wss"
-	} else {
-		u.Scheme = "ws"
-	}
-
-	qs := u.Query()
-	qs.Set("ws", "true")
-	qs["command"] = args.Command
-	qs.Set("pod", args.Pod)
-	qs.Set("container", args.Container)
-	qs.Set("interactive", strconv.FormatBool(args.Interactive))
-	qs.Set("tty", strconv.FormatBool(args.TTY))
-	qs.Set("width", strconv.FormatInt(int64(args.TerminalWidth), 10))
-	qs.Set("height", strconv.FormatInt(int64(args.TerminalHeight), 10))
-
-	u.RawQuery = qs.Encode()
-
-	conn, _, err := c.ws.DialContext(ctx, u.String(), c.baseAuthHeader(nil))
-	if err != nil {
-		return nil, err
-	}
-
-	if args.In != nil {
-		go io.Copy(&wsWriter{conn}, args.In)
-	}
-
-	return conn, nil
-}
-
-type wsWriter struct {
-	*websocket.Conn
-}
-
-func (w *wsWriter) Write(p []byte) (int, error) {
-	return len(p), w.Conn.WriteMessage(websocket.BinaryMessage, p)
-}
-
-func (c *client) buildRequest(operation string, data interface{}) (req *http.Request, err error) {
-	switch operation {
-	case "Scale":
-		args := data.(ScaleArgs)
-		pathName := fmt.Sprintf("/resources/%s/scale", args.Instance)
-		values := url.Values{}
-		values.Set("quantity", fmt.Sprint(args.Replicas))
-		body := strings.NewReader(values.Encode())
-		req, err = c.newRequest("POST", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-	case "Info":
-		args := data.(InfoArgs)
-		pathName := fmt.Sprintf("/resources/%s/info", args.Instance)
-		req, err = c.newRequest("GET", pathName, nil, args.Instance)
-
-	case "UpdateCertificate":
-		args := data.(UpdateCertificateArgs)
-		buffer := &bytes.Buffer{}
-		w := multipart.NewWriter(buffer)
-
-		if args.boundary != "" {
-			if err = w.SetBoundary(args.boundary); err != nil {
-				return nil, err
-			}
-		}
-
-		var part io.Writer
-		{
-			part, err = w.CreateFormFile("cert", "cert.pem")
-			if err != nil {
-				return nil, err
-			}
-
-			part.Write([]byte(args.Certificate))
-		}
-		{
-			part, err = w.CreateFormFile("key", "key.pem")
-			if err != nil {
-				return nil, err
-			}
-
-			part.Write([]byte(args.Key))
-		}
-
-		if err = w.WriteField("name", args.Name); err != nil {
-			return nil, err
-		}
-
-		if err = w.Close(); err != nil {
-			return nil, err
-		}
-
-		body := strings.NewReader(buffer.String())
-		pathName := fmt.Sprintf("/resources/%s/certificate", args.Instance)
-		req, err = c.newRequest("POST", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", fmt.Sprintf("multipart/form-data; boundary=%q", w.Boundary()))
-
-	case "DeleteCertificate":
-		args := data.(DeleteCertificateArgs)
-		pathName := fmt.Sprintf("/resources/%s/certificate/%s", args.Instance, args.Name)
-		req, err = c.newRequest("DELETE", pathName, nil, args.Instance)
-
-	case "UpdateBlock":
-		args := data.(UpdateBlockArgs)
-		values := url.Values{}
-		values.Set("block_name", args.Name)
-		values.Set("content", args.Content)
-		body := strings.NewReader(values.Encode())
-		pathName := fmt.Sprintf("/resources/%s/block", args.Instance)
-		req, err = c.newRequest("POST", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-	case "DeleteBlock":
-		args := data.(DeleteBlockArgs)
-		pathName := fmt.Sprintf("/resources/%s/block/%s", args.Instance, args.Name)
-		req, err = c.newRequest("DELETE", pathName, nil, args.Instance)
-
-	case "ListBlocks":
-		args := data.(ListBlocksArgs)
-		pathName := fmt.Sprintf("/resources/%s/block", args.Instance)
-		req, err = c.newRequest("GET", pathName, nil, args.Instance)
-
-	case "GetAutoscale":
-		args := data.(GetAutoscaleArgs)
-		pathName := fmt.Sprintf("/resources/%s/autoscale", args.Instance)
-		req, err = c.newRequest("GET", pathName, nil, args.Instance)
-
-	case "RemoveAutoscale":
-		args := data.(RemoveAutoscaleArgs)
-		pathName := fmt.Sprintf("/resources/%s/autoscale", args.Instance)
-		req, err = c.newRequest("DELETE", pathName, nil, args.Instance)
-
-	case "CreateAutoscale":
-		args := data.(UpdateAutoscaleArgs)
-		pathName := fmt.Sprintf("/resources/%s/autoscale", args.Instance)
-		values := url.Values{}
-		values.Set("max", fmt.Sprint(args.MaxReplicas))
-		values.Set("min", fmt.Sprint(args.MinReplicas))
-		values.Set("cpu", fmt.Sprint(args.CPU))
-		values.Set("memory", fmt.Sprint(args.Memory))
-		body := strings.NewReader(values.Encode())
-		req, err = c.newRequest("POST", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-	case "UpdateAutoscale":
-		args := data.(UpdateAutoscaleArgs)
-		pathName := fmt.Sprintf("/resources/%s/autoscale", args.Instance)
-		values := url.Values{}
-		if args.MaxReplicas > 0 {
-			values.Set("max", fmt.Sprint(args.MaxReplicas))
-		}
-		if args.MinReplicas > 0 {
-			values.Set("min", fmt.Sprint(args.MinReplicas))
-		}
-		if args.CPU > 0 {
-			values.Set("cpu", fmt.Sprint(args.CPU))
-		}
-		if args.Memory > 0 {
-			values.Set("memory", fmt.Sprint(args.Memory))
-		}
-		body := strings.NewReader(values.Encode())
-		req, err = c.newRequest("PATCH", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-	case "DeleteRoute":
-		args := data.(DeleteRouteArgs)
-		pathName := fmt.Sprintf("/resources/%s/route", args.Instance)
-		values := url.Values{}
-		values.Set("path", args.Path)
-		body := strings.NewReader(values.Encode())
-		req, err = c.newRequest("DELETE", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-	case "ListRoutes":
-		args := data.(ListRoutesArgs)
-		pathName := fmt.Sprintf("/resources/%s/route", args.Instance)
-		req, err = c.newRequest("GET", pathName, nil, args.Instance)
-
-	case "UpdateRoute":
-		args := data.(UpdateRouteArgs)
-		pathName := fmt.Sprintf("/resources/%s/route", args.Instance)
-		values := url.Values{
-			"path":        []string{args.Path},
-			"destination": []string{args.Destination},
-			"https_only":  []string{strconv.FormatBool(args.HTTPSOnly)},
-			"content":     []string{args.Content},
-		}
-		body := strings.NewReader(values.Encode())
-		req, err = c.newRequest("POST", pathName, body, args.Instance)
-		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-	default:
-		err = fmt.Errorf("rpaasv2: unknown operation")
-	}
-
-	return
 }
 
 func (c *client) newRequest(method, pathName string, body io.Reader, instance string) (*http.Request, error) {
