+++ pkg/rpaas/client/internal_client_test.go
@@ -5,13 +5,9 @@
 package client
 
 import (
-	"context"
-	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
 	"os"
 	"testing"
 	"time"
@@ -19,7 +15,6 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/tsuru/rpaas-operator/pkg/rpaas/client/types"
 )
 
 var (
@@ -127,940 +122,6 @@ func TestNewClientThroughTsuruWithOptions(t *testing.T) {
 			if tt.teardown != nil {
 				tt.teardown(t)
 			}
-		})
-	}
-}
-
-func TestClientThroughTsuru_Scale(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          ScaleArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name: "when replicas number is negative",
-			args: ScaleArgs{
-				Instance: "my-instance",
-				Replicas: int32(-1),
-			},
-			expectedError: "rpaasv2: replicas must be greater or equal than zero",
-		},
-		{
-			name: "when server returns an unexpected status code",
-			args: ScaleArgs{
-				Instance: "my-instance",
-				Replicas: int32(777),
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when server returns the expected response",
-			args: ScaleArgs{
-				Instance: "my-instance",
-				Replicas: int32(777),
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "POST")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/scale"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				assert.Equal(t, "application/x-www-form-urlencoded", r.Header.Get("Content-Type"))
-				assert.Equal(t, "quantity=777", getBody(t, r))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.Scale(context.TODO(), tt.args)
-			if tt.expectedError == "" {
-				require.NoError(t, err)
-				return
-			}
-			assert.EqualError(t, err, tt.expectedError)
-		})
-	}
-}
-
-func TestClientThroughTsuru_UpdateCertificate(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          UpdateCertificateArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name: "when instance is empty",
-			args: UpdateCertificateArgs{
-				Certificate: "some cert",
-				Key:         "some key",
-			},
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when certificate is empty",
-			args: UpdateCertificateArgs{
-				Instance: "my-instance",
-				Key:      "some key",
-			},
-			expectedError: "rpaasv2: certificate cannot be empty",
-		},
-		{
-			name: "when key is empty",
-			args: UpdateCertificateArgs{
-				Instance:    "my-instance",
-				Certificate: "some cert",
-			},
-			expectedError: "rpaasv2: key cannot be empty",
-		},
-		{
-			name: "when the server returns the expected response",
-			args: UpdateCertificateArgs{
-				Instance:    "my-instance",
-				Certificate: `my certificate`,
-				Key:         `my key`,
-				boundary:    "custom-boundary",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "POST")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/certificate"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				assert.Equal(t, "multipart/form-data; boundary=\"custom-boundary\"", r.Header.Get("Content-Type"))
-				assert.Equal(t, "--custom-boundary\r\nContent-Disposition: form-data; name=\"cert\"; filename=\"cert.pem\"\r\nContent-Type: application/octet-stream\r\n\r\nmy certificate\r\n--custom-boundary\r\nContent-Disposition: form-data; name=\"key\"; filename=\"key.pem\"\r\nContent-Type: application/octet-stream\r\n\r\nmy key\r\n--custom-boundary\r\nContent-Disposition: form-data; name=\"name\"\r\n\r\n\r\n--custom-boundary--\r\n", getBody(t, r))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-		{
-			name: "when the server returns an error",
-			args: UpdateCertificateArgs{
-				Instance:    "my-instance",
-				Certificate: `my certificate`,
-				Key:         `my key`,
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.UpdateCertificate(context.TODO(), tt.args)
-			if tt.expectedError == "" {
-				require.NoError(t, err)
-				return
-			}
-			assert.EqualError(t, err, tt.expectedError)
-		})
-	}
-}
-
-func TestClientThroughTsuru_DeleteCertificate(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          DeleteCertificateArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			args:          DeleteCertificateArgs{},
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when certificate name is empty, should not return error since empty = default",
-			args: DeleteCertificateArgs{
-				Instance: "my-instance",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "DELETE")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/certificate/"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: DeleteCertificateArgs{
-				Instance: "my-instance",
-				Name:     "my-certificate",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "DELETE")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/certificate/my-certificate"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				assert.Equal(t, "", getBody(t, r))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-		{
-			name: "when the server returns an error",
-			args: DeleteCertificateArgs{
-				Instance: "my-instance",
-				Name:     "my-certificate",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the certificate name has spaces should should query escape cert name",
-			args: DeleteCertificateArgs{
-				Instance: "my-instance",
-				Name:     "my certificate",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/certificate/my+certificate"), r.URL.RequestURI())
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.DeleteCertificate(context.TODO(), tt.args)
-			if tt.expectedError == "" {
-				require.NoError(t, err)
-				return
-			}
-			assert.EqualError(t, err, tt.expectedError)
-		})
-	}
-}
-
-func TestClientThroughTsuru_UpdateBlock(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          UpdateBlockArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when block name is empty",
-			args: UpdateBlockArgs{
-				Instance: "some-instance",
-				Content:  "some content",
-			},
-			expectedError: "rpaasv2: block name cannot be empty",
-		},
-		{
-			name: "when content is empty",
-			args: UpdateBlockArgs{
-				Instance: "my-instance",
-				Name:     "server",
-			},
-			expectedError: "rpaasv2: content cannot be empty",
-		},
-		{
-			name: "when the server returns the expected response",
-			args: UpdateBlockArgs{
-				Instance: "my-instance",
-				Name:     "http",
-				Content:  "# NGINX configuration block",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "POST")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/block"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				assert.Equal(t, "application/x-www-form-urlencoded", r.Header.Get("Content-Type"))
-				assert.Equal(t, "block_name=http&content=%23+NGINX+configuration+block", getBody(t, r))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-		{
-			name: "when the server returns an error",
-			args: UpdateBlockArgs{
-				Instance: "my-instance",
-				Name:     "server",
-				Content:  "Some NGINX snippet",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.UpdateBlock(context.TODO(), tt.args)
-			if tt.expectedError == "" {
-				require.NoError(t, err)
-				return
-			}
-			assert.EqualError(t, err, tt.expectedError)
-		})
-	}
-}
-
-func TestClientThroughTsuru_DeleteBlock(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          DeleteBlockArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when block name is empty",
-			args: DeleteBlockArgs{
-				Instance: "some-instance",
-			},
-			expectedError: "rpaasv2: block name cannot be empty",
-		},
-		{
-			name: "when the server returns the expected response",
-			args: DeleteBlockArgs{
-				Instance: "my-instance",
-				Name:     "http",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "DELETE")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/block/http"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-		{
-			name: "when the server returns an error",
-			args: DeleteBlockArgs{
-				Instance: "my-instance",
-				Name:     "server",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.DeleteBlock(context.TODO(), tt.args)
-			if tt.expectedError == "" {
-				require.NoError(t, err)
-				return
-			}
-			assert.EqualError(t, err, tt.expectedError)
-		})
-	}
-}
-
-func TestClientThroughTsuru_ListBlocks(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          ListBlocksArgs
-		expected      []types.Block
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: ListBlocksArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: ListBlocksArgs{
-				Instance: "my-instance",
-			},
-			expected: []types.Block{
-				{
-					Name:    "http",
-					Content: "Some HTTP conf",
-				},
-				{
-					Name:    "server",
-					Content: "Some server conf",
-				},
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "GET")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/block"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				fmt.Fprintf(w, `{"blocks": [{"block_name": "http", "content": "Some HTTP conf"}, {"block_name": "server", "content": "Some server conf"}]}`)
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			blocks, err := client.ListBlocks(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-			assert.Equal(t, tt.expected, blocks)
-		})
-	}
-}
-
-func TestClientThroughTsuru_DeleteRoute(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          DeleteRouteArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when path is empty",
-			args: DeleteRouteArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: path cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: DeleteRouteArgs{
-				Instance: "my-instance",
-				Path:     "/custom/path",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: DeleteRouteArgs{
-				Instance: "my-instance",
-				Path:     "/custom/path",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "DELETE")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/route"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				assert.Equal(t, "application/x-www-form-urlencoded", r.Header.Get("Content-Type"))
-				assert.Equal(t, "path=%2Fcustom%2Fpath", getBody(t, r))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.DeleteRoute(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestClientThroughTsuru_ListRoutes(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          ListRoutesArgs
-		expected      []types.Route
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: ListRoutesArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: ListRoutesArgs{
-				Instance: "my-instance",
-			},
-			expected: []types.Route{
-				{
-					Path:        "/static",
-					Destination: "static.apps.tsuru.example.com",
-				},
-				{
-					Path:        "/login",
-					Destination: "login.apps.tsuru.example.com",
-					HTTPSOnly:   true,
-				},
-				{
-					Path:    "/custom/path",
-					Content: "# some NGINX configuration",
-				},
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "GET")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/route"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				fmt.Fprintf(w, `{"paths": [{"path": "/static", "destination": "static.apps.tsuru.example.com"}, {"path": "/login", "destination": "login.apps.tsuru.example.com", "https_only": true}, {"path": "/custom/path", "content": "# some NGINX configuration"}]}`)
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			blocks, err := client.ListRoutes(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-			assert.Equal(t, tt.expected, blocks)
-		})
-	}
-}
-
-func TestClientThroughTsuru_Info(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          InfoArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when all args are valid",
-			args: InfoArgs{
-				Instance: "my-instance",
-				Raw:      true,
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "GET")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/info"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				fmt.Fprintf(w, `{"address":[{"hostname":"some-host","ip":"0.0.0.0"},{"hostname":"some-host2","ip":"0.0.0.1"}],"replicas":5,"plan":"basic","locations":[{"path":"some-path","destination":"some-destination"}],"binds":[{"name":"some-name","host":"some-host"},{"name":"some-name2","host":"some-host2"}],"team":"some team","name":"my-instance","description":"some description","tags":["tag1","tag2","tag3"]}`)
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			_, err := client.Info(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestClientThroughTsuru_UpdateRoute(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          UpdateRouteArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when path is empty",
-			args: UpdateRouteArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: path cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: UpdateRouteArgs{
-				Instance:    "my-instance",
-				Path:        "/app",
-				Destination: "app.tsuru.example.com",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: UpdateRouteArgs{
-				Instance:    "my-instance",
-				Path:        "/app",
-				Destination: "app.tsuru.example.com",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "POST")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/route"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				expected := url.Values{
-					"path":        []string{"/app"},
-					"destination": []string{"app.tsuru.example.com"},
-					"https_only":  []string{"false"},
-					"content":     []string{""},
-				}
-				values, err := url.ParseQuery(getBody(t, r))
-				assert.NoError(t, err)
-				assert.Equal(t, expected, values)
-				w.WriteHeader(http.StatusCreated)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.UpdateRoute(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestClientThroughTsuru_UpdateAutoscale(t *testing.T) {
-	handlerCount := 0
-	tests := []struct {
-		name          string
-		args          UpdateAutoscaleArgs
-		expectedError string
-		handler       http.HandlerFunc
-		handlerCount  int
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: UpdateAutoscaleArgs{
-				Instance:    "my-instance",
-				MinReplicas: int32(5),
-				MaxReplicas: int32(10),
-				CPU:         int32(27),
-				Memory:      int32(33),
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when an autoscale spec is found",
-			args: UpdateAutoscaleArgs{
-				Instance:    "my-instance",
-				MinReplicas: int32(5),
-				MaxReplicas: int32(10),
-				CPU:         int32(27),
-				Memory:      int32(33),
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				switch handlerCount {
-				case 0:
-					assert.Equal(t, "GET", r.Method)
-					assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-					assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/autoscale"), r.URL.RequestURI())
-					w.WriteHeader(http.StatusOK)
-					payload, err := json.Marshal(&types.Autoscale{})
-					require.NoError(t, err)
-					w.Write(payload)
-					handlerCount++
-
-				case 1:
-					assert.Equal(t, "PATCH", r.Method)
-					assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-					assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/autoscale"), r.URL.RequestURI())
-					expected := url.Values{
-						"max":    []string{"10"},
-						"min":    []string{"5"},
-						"cpu":    []string{"27"},
-						"memory": []string{"33"},
-					}
-					values, err := url.ParseQuery(getBody(t, r))
-					assert.NoError(t, err)
-					assert.Equal(t, expected, values)
-					w.WriteHeader(http.StatusCreated)
-					handlerCount = 0
-				}
-			},
-		},
-		{
-			name: "when an autoscale spec is not found",
-			args: UpdateAutoscaleArgs{
-				Instance:    "my-instance",
-				MinReplicas: int32(5),
-				MaxReplicas: int32(10),
-				CPU:         int32(27),
-				Memory:      int32(33),
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				switch handlerCount {
-				case 0:
-					assert.Equal(t, "GET", r.Method)
-					assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-					assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/autoscale"), r.URL.RequestURI())
-					w.WriteHeader(http.StatusNotFound)
-					handlerCount++
-
-				case 1:
-					assert.Equal(t, "POST", r.Method)
-					assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-					assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/autoscale"), r.URL.RequestURI())
-					expected := url.Values{
-						"max":    []string{"10"},
-						"min":    []string{"5"},
-						"cpu":    []string{"27"},
-						"memory": []string{"33"},
-					}
-					values, err := url.ParseQuery(getBody(t, r))
-					assert.NoError(t, err)
-					assert.Equal(t, expected, values)
-					w.WriteHeader(http.StatusCreated)
-					handlerCount = 0
-				}
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.UpdateAutoscale(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestClientThroughTsuru_GetAutoscale(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          GetAutoscaleArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: GetAutoscaleArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: GetAutoscaleArgs{
-				Instance: "my-instance",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "GET")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/autoscale"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				w.WriteHeader(http.StatusOK)
-				fmt.Fprintf(w, "{\n\t\"minReplicas\": 2,\n\t\"maxReplicas\": 5,\n\t\"cpu\": 50,\n\t\"memory\": 55\n}\n")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			_, err := client.GetAutoscale(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestClientThroughTsuru_RemoveAutoscale(t *testing.T) {
-	tests := []struct {
-		name          string
-		args          RemoveAutoscaleArgs
-		expectedError string
-		handler       http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when the server returns an error",
-			args: RemoveAutoscaleArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: unexpected status code: 404 Not Found, detail: instance not found",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusNotFound)
-				fmt.Fprintf(w, "instance not found")
-			},
-		},
-		{
-			name: "when the server returns the expected response",
-			args: RemoveAutoscaleArgs{
-				Instance: "my-instance",
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				assert.Equal(t, r.Method, "DELETE")
-				assert.Equal(t, fmt.Sprintf("/services/%s/proxy/%s?callback=%s", FakeTsuruService, "my-instance", "/resources/my-instance/autoscale"), r.URL.RequestURI())
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				w.WriteHeader(http.StatusOK)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			err := client.RemoveAutoscale(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestClientThroughTsuru_Exec(t *testing.T) {
-	var called bool
-	tests := []struct {
-		name           string
-		args           ExecArgs
-		expectedError  string
-		expectedCalled bool
-		handler        http.HandlerFunc
-	}{
-		{
-			name:          "when instance is empty",
-			expectedError: "rpaasv2: instance cannot be empty",
-		},
-		{
-			name: "when command is not set",
-			args: ExecArgs{
-				Instance: "my-instance",
-			},
-			expectedError: "rpaasv2: command cannot be empty",
-		},
-		{
-			name: "when all options are set",
-			args: ExecArgs{
-				Instance:       "my-instance",
-				Command:        []string{"bash"},
-				Pod:            "pod-1",
-				Container:      "nginx",
-				Interactive:    true,
-				TTY:            true,
-				TerminalWidth:  uint16(80),
-				TerminalHeight: uint16(24),
-			},
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				called = true
-				assert.True(t, websocket.IsWebSocketUpgrade(r))
-				assert.Equal(t, "Bearer f4k3t0k3n", r.Header.Get("Authorization"))
-				expectedQS := url.Values{}
-				expectedQS.Set("callback", "/resources/my-instance/exec")
-				expectedQS.Set("ws", "true")
-				expectedQS.Set("command", "bash")
-				expectedQS.Set("pod", "pod-1")
-				expectedQS.Set("container", "nginx")
-				expectedQS.Set("width", "80")
-				expectedQS.Set("height", "24")
-				expectedQS.Set("tty", "true")
-				expectedQS.Set("interactive", "true")
-				assert.Equal(t, expectedQS, r.URL.Query())
-				w.WriteHeader(http.StatusBadRequest)
-			},
-			expectedCalled: true,
-			expectedError:  "websocket: bad handshake",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			called = false
-			client, server := newClientThroughTsuru(t, tt.handler)
-			defer server.Close()
-			_, err := client.Exec(context.TODO(), tt.args)
-			if tt.expectedError != "" {
-				assert.EqualError(t, err, tt.expectedError)
-				return
-			}
-			assert.NoError(t, err)
-			assert.Equal(t, tt.expectedCalled, called)
 		})
 	}
 }
